refactor and DRY code
Extracted common parts from both userscript and extension into a single
javascript file, and inserted `//=INSERT_FILE=` placeholders instead.
This is meant to be used as template files.

Also the inject CSS style was extracted and replaced with an instruction
`//=INSERT_FILE_AS_STRINGS=` which will include the CSS file instead.
This is handled differently from previous (insert file lines as quoted
strings which are then merged into one big string).

A ruby rakefile is used to process the source files and insert the
imported files where specified.

The sources are placed in the `src` directory, and the final built files
are created in `js` directory.